Documentation for ongoing work (#1)
* docs: Add progress document for cairo-components


* Refactor: Document variable log size support in recursive circuits


* Checkpoint before follow-up message


* Add CI workflow for Rust projects


* Refactor: Improve code quality and fix bugs

This commit includes various refactors and bug fixes across multiple modules. Key changes include:

- **Code Quality Improvements**:
    - Replaced `clone()` with direct usage where appropriate.
    - Simplified variable assignments and destructuring.
    - Used iterators more effectively.
    - Improved error handling and removed unnecessary unwraps.
    - Updated dependencies and removed unused code.

- **Bug Fixes**:
    - Corrected logic in hashing and Merkle tree operations.
    - Fixed issues related to variable selection and conditional logic.
    - Addressed potential off-by-one errors in array indexing and loops.

These changes aim to enhance the overall robustness and maintainability of the codebase.


* fix clippy

* keep the original decommitment

* fmt

* push a fix

* clippy

* fmt

---------

+++ cairo-components/recursive/data_structures/src/claim.rs
@@ -349,11 +349,11 @@ impl AllocVar for CairoClaimVar {
         let public_data = PublicDataVar::new_variables(cs, &value.public_data, mode);
         let opcode_claim = OpcodeClaimVar::new_variables(cs, &value.opcodes, mode);
         let verify_instruction =
-            LogSizeVar::new_variables(cs, &(value.verify_instruction.log_size as u32), mode);
+            LogSizeVar::new_variables(cs, &value.verify_instruction.log_size, mode);
         let blake_context = BlakeContextClaimVar::new_variables(cs, &value.blake_context, mode);
         let builtins = BuiltinsClaimVar::new_variables(cs, &value.builtins, mode);
         let memory_address_to_id =
-            LogSizeVar::new_variables(cs, &(value.memory_address_to_id.log_size as u32), mode);
+            LogSizeVar::new_variables(cs, &value.memory_address_to_id.log_size, mode);
         let memory_id_to_value =
             MemoryIdToBigClaimVar::new_variables(cs, &value.memory_id_to_value, mode);
         Self {
